Improve column mask calculation
If the update operation changes a field with number >= 64, the
column mask of the update op is set to UINT64_MAX.
Lets use the last bit of the column mask as the flag, that all
fields with numbers >= 63 could be changed. Then if the indexed
positions are less than 64, the column mask will be always working.

Closes #1716