vinyl: fix vy_index leak on index creation error path
VinylIndex::open() doesn't free vy_index if vy_index_open() fails.
Also, VinylIndex leaks vy_index if we fail to log index creation in WAL,
because in this case VinylIndex::commitDrop(), which releases the
underlying vy_index, isn't called. To fix the leaks, make VinylIndex
reference vy_index object which it is wrapped around in constructor and
unreference it in destructor.

While we are at it, remove useless TODOs from vy_index_open() and
vy_index_commit_drop() asking not to drop and recreate index on
recovery, because we don't reload index directory on drop/create since
the time vylog was introduced.