vinyl: log index creation after wal write
Currently, we log index creation before writing the index row to WAL,
via VinylEngine::addPrimaryKey() and VinylEngine::buildSecondaryKey().
This is incorrect, because it isn't guaranteed that WAL write will
succeed. If it doesn't, the index creation is aborted, but the index
record is left in the metadata log. At best this will result in garbage
in the metadata log, at worst we will get a permanent failure to rotate
the log. The latter can happen if the index id in the metadata log (i.e.
LSN) is reused by another index, e.g. after the following piece of Lua
code is executed by tarantool, box.snapshot() will keep failing:

    tarantool> box.cfg{}
    ---
    ...

    tarantool> s = box.schema.space.create("test", {engine='vinyl'})
    ---
    ...

    tarantool> box.error.injection.set('ERRINJ_WAL_IO', true)
    ---
    - ok
    ...

    tarantool> _ = box.space.test:create_index('pk')
    ---
    - error: Failed to write to disk
    ...

    tarantool> box.error.injection.set('ERRINJ_WAL_IO', false)
    ---
    - ok
    ...

    tarantool> _ = box.space.test:create_index('pk')
    ---
    ...

    tarantool> box.snapshot()
    ---
    - error: 'Invalid VYLOG file: Duplicate index id 2'
    ...

Fix this by moving vinyl index creation logging to the new Index method
commitCreate(), which is called after WAL write. Since we can't fail
there, we use the same technique as in commitDrop() to make sure the log
record is committed sooner or later, namely in case of vylog write
failure we leave the record in the vylog buffer to be flushed along with
the next write. If it doesn't get flushed before shutdown, we will
replay it on local recovery from WAL.